requirements.txt
streamlit>=1.37
pandas>=2.2
pydantic>=2.6
python-dotenv>=1.0
requests>=2.32
